backend-python/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
aiosqlite==0.19.0
requests==2.31.0
openai>=1.3.0
python-dotenv==1.0.0
pytest==7.4.3
httpx>=0.26.0
reportlab>=4.1.0
